Move inheritance-related tests
These tests are interesting and important for polymorphism work, move them to their own source file
